requirements.txt
requests
pygerduty
redis
pyyaml
